+++ Diploma/factors.cpp
@@ -6,53 +6,53 @@
 #include <cmath>
 
 namespace ftr {
-    static double const moveVelocity = 0.75 / 60; // м/с
+    static float const moveVelocity = 0.75 / 60; // м/с
 
-    static double const TLik = 1738; // К
-    static double const TSol = 1679; // К
-    static double const dT = 0.000001; // K ???
+    static float const TLik = 1738; // К
+    static float const TSol = 1679; // К
+    static float const dT = 0.000001; // K ???
 
-    static double const L = 272; // кДж/кг
-    static double const cLik = 710; // Дж/(кг * К)
-    static double const x = 0.7;
+    static float const L = 272; // кДж/кг
+    static float const cLik = 710; // Дж/(кг * К)
+    static float const x = 0.7;
 
-    static double const totalLength = (0.4 + 0.4 + 0.47 + 0.95 + 1.51 + 18.97); // с
+    static float const totalLength = (0.4 + 0.4 + 0.47 + 0.95 + 1.51 + 18.97); // с
 
-    static double Temps[] = { 273, 373, 473, 573, 673, 773, 873,
+    static float Temps[] = { 273, 373, 473, 573, 673, 773, 873,
         973, 1073, 1173, 1273, 1373, 1473, 1679, 1682, 1800
     };
-    static double Lambds[] = {
+    static float Lambds[] = {
         52.56057, 51.35258, 49.16971, 46.22939, 42.74907, 38.94618, 35.03819,
         31.24254, 24.06517, 25.37233, 26.95363, 28.32515, 29.40302, 31.62343,
         28.0, 28.0
     };
-    static double Ros[] = {
+    static float Ros[] = {
         7885.884, 7845.138, 7804.392, 7763.646, 7722.9, 7682.901, 7647.512, 7621.141,
         7631.934, 7572.359, 7512.151, 7453.459, 7398.322, 7190.562, 7000.0, 7000.0,
     };
 
-    static double Ti[] = { 1000, 1033, 923, 1033 };
-    static double dTi[] = { 70, 350, 1100, 170 };
+    static float Ti[] = { 1000, 1033, 923, 1033 };
+    static float dTi[] = { 70, 350, 1100, 170 };
 
-    inline double alphaForT(double T, size_t &index) {
+    inline float alphaForT(float T, size_t &index) {
         index = 0;
         while (Temps[index] < T) ++index;
         return (T - Temps[index - 1]) / (Temps[index] - Temps[index - 1]);
     }
 
-    inline double Lambda(double T) {
+    inline float Lambda(float T) {
         size_t index = 0;
-        double alpha = alphaForT(T, index);
+        float alpha = alphaForT(T, index);
         return Lambds[index - 1] + alpha * (Lambds[index] - Lambds[index - 1]);
     }
 
-    inline double Ro(double T) {
+    inline float Ro(float T) {
         size_t index = 0;
-        double alpha = alphaForT(T, index);
+        float alpha = alphaForT(T, index);
         return Ros[index - 1] + alpha * (Ros[index] - Ros[index - 1]);
     }
 
-    inline double Li(unsigned short i, double x) {
+    inline float Li(unsigned short i, float x) {
         switch (i) {
             case 0:
                 return 44076 - 85622 * x * x + 50357 * x;
@@ -66,23 +66,23 @@ namespace ftr {
         return 0.0;
     }
 
-    inline double cSol(double T) {
-        double result = 469 + 0.16 * (T - 323);
+    inline float cSol(float T) {
+        float result = 469 + 0.16 * (T - 323);
         for (unsigned short i = 0; i < 4; ++i) {
-            double tC = (Ti[i] - T) / dTi[i];
+            float tC = (Ti[i] - T) / dTi[i];
             result += 4.5141 * Li(i, x) / dTi[i] * exp(-16 * tC * tC);
         }
         return result;
     }
 
-    inline double sigm(double T) {
-        double cLikS = 15.463359 - 0.124528e-1 * T + 0.216279e-5 * T * T;
-        double cSolS = -11.0388 + 0.278656e-1 * T - 0.120163e-4 * T * T;
+    inline float sigm(float T) {
+        float cLikS = 15.463359 - 0.124528e-1 * T + 0.216279e-5 * T * T;
+        float cSolS = -11.0388 + 0.278656e-1 * T - 0.120163e-4 * T * T;
         return (cLikS - 0.7) / (cLikS - cSolS);
     }
 }
 
-double Factors::cEf(double T) const {
+float Factors::cEf(float T) const {
     if (T >= ftr::TLik) {
         return ftr::cLik;
     }
@@ -94,8 +94,8 @@ double Factors::cEf(double T) const {
     }
 }
 
-double Factors::alpha(double t) const {
-    double x = t * ftr::moveVelocity;
+float Factors::alpha(float t) const {
+    float x = t * ftr::moveVelocity;
 
     if (x <= 0.4) {
         return 2100;
@@ -117,8 +117,8 @@ double Factors::alpha(double t) const {
     }
 }
 
-double Factors::sigma(double t) const {
-    double x = t * ftr::moveVelocity;
+float Factors::sigma(float t) const {
+    float x = t * ftr::moveVelocity;
 
     if (x <= 0.4 + 0.4 + 0.47 + 0.95 + 1.51) {
         return 0;
@@ -128,52 +128,52 @@ double Factors::sigma(double t) const {
     }
 }
 
-double Factors::lambda(double T) const {
+float Factors::lambda(float T) const {
     return ftr::Lambda(T);
 }
 
-double Factors::ro(double T) const {
+float Factors::ro(float T) const {
     return ftr::Ro(T);
 }
 
-double Factors::X1() const {
+float Factors::X1() const {
     return config.value("X1");
 }
 
-double Factors::X2() const {
+float Factors::X2() const {
     return config.value("X2");
 }
 
-double Factors::totalTime() const {
+float Factors::totalTime() const {
     return ftr::totalLength / config.value("Speed");
 }
 
-double Factors::X1SplitCount() const {
+float Factors::X1SplitCount() const {
     return config.value("X1SplitCount");
 }
 
-double Factors::X2SplitCount() const {
+float Factors::X2SplitCount() const {
     return config.value("X2SplitCount");
 }
 
-double Factors::TimeSplitCount() const {
+float Factors::TimeSplitCount() const {
     return config.value("TimeSplitCount");
 }
 
-double Factors::Epsilon() const {
+float Factors::Epsilon() const {
     return config.value("Epsilon");
 }
 
-double Factors::TStart() const {
+float Factors::TStart() const {
     return config.value("InitT");
 }
 
-double Factors::TEnv() const {
+float Factors::TEnv() const {
     return config.value("EnvT");
 }
 
-double Factors::TEnv4() const {
-    double value = TEnv();
+float Factors::TEnv4() const {
+    float value = TEnv();
     value *= value;
     value *= value;
     return value;
@@ -203,7 +203,7 @@ size_t Factors::MatrixFramesCount() const {
     return config.value("MatrixFramesCount");
 }
 
-double Factors::X1View(size_t index) const {
+float Factors::X1View(size_t index) const {
     char buff[10];
     snprintf(buff, 10, "View%luX1", index);
     std::string key = buff;
@@ -211,7 +211,7 @@ double Factors::X1View(size_t index) const {
     return config.value(key);
 }
 
-double Factors::X2View(size_t index) const {
+float Factors::X2View(size_t index) const {
     char buff[10];
     snprintf(buff, 10, "View%luX2", index);
     std::string key = buff;
